Changed int to unsigned long, since .capacity() gives unsigned long.

+++ include/exchange.h
@@ -25,7 +25,7 @@ public:
   virtual bool subscribed_to_OHLC(std::chrono::minutes) = 0;
   virtual void userinfo() = 0;
   virtual void ping() = 0;
-  virtual bool backfill_OHLC(std::chrono::minutes, int) = 0;
+  virtual bool backfill_OHLC(std::chrono::minutes, unsigned long) = 0;
   virtual std::string status() = 0;
   virtual bool connected() = 0;
